+++ tests/Fixture/ref.php
@@ -7,8 +7,8 @@ namespace Thesis\Fixture;
 use Thesis\Dic\Configuration\ValueConfig;
 use Thesis\Dic\Internal\Autowiring;
 use Thesis\Dic\Internal\Builder;
-use Thesis\Dic\Location;
 use Thesis\Dic\Ref;
+use function Thesis\Dic\Internal\caller;
 
 /**
  * @template T
@@ -21,6 +21,6 @@ function ref(mixed $value = null): Ref
         builder: new Builder(),
         autowiring: new Autowiring(),
         value: $value,
-        declaredAt: Location::caller(),
+        declaredAt: caller(),
     );
 }